test for missing type warnings
